Add initial anchorhub.py file - basics working
anchorhub.py is the initial, unrefactored, not-structured-that-well file
that has the first pieces of functionality built in and working for the
most part. It is capable of taking a directory tree, looking through
each of the markdown files, and modifying the headers, internal anchor
links, and external anchor links (those that reference other files'
marked anchors). It also supports arbitrary anchor-wrappers, over-riding
files, and outputting to different locations.

There is still plenty to get done! In no particular order:

* Command line argument support
* Support for Markdown reference links
* Turn it into something that is pip install-able
* Ensure cross-platform compatibility (only testing on OSX right now)
* Proper exception handling
* Refactor and reorganize
* Throw more tests at it