Replace anchor extraction with AnchorCollector
Instead of hard coding the first pass through the files, anchors and
converted header lines are parsed using an AnchorCollector, an object
that is designed to parse through files using given strategies. The
strategies used here are GitHub and Markdown style, which will be the
default. Ideally, the API will be clean and easy enough to use for
various types of files and auto-generated header formats.

File writing will have its own object-strategy structure, and will
hopefully be done soon!