Fix cmdparse bug, and argument to 'extensions'
Directly passing in sys.args to parse_args() is different than calling
parse_args() without any parameters. When you directly pass it in, it
does not ignore the default value of sys.args[0], which is the path of
the executed file. Tweaked this to make it work correctly.

Additionally, all validators have used the attribute 'extensions'
instead of the attribute 'extension' in their code, and I think that
makes more sense. Changing the command-line input to match this.