Add get_label() functionality to WriterStrategy
Instead of having the label be fixed for each strategy, the labels will
be given to the concrete WriterStrategy objects upon instantiation. The
abstract super class, WriterStrategy implements the assignment of the
label to a local variable in its __init__() method. It also implements
the get_label() getter, so child WriterStrategy classes will not need to
implement it.

The super().__init__() method should be called inside of concrete
WriterStrategy __init__() methods.