Add duplicate-anchor error functionality
Before, the program was detecting and collecting duplicate AnchorHub
tags specified within the same file, but nothing was being done with
them. This adds an extra assertion to in main.py to validate the
collected anchors. If the duplicate_tags dictionary is not empty,
AnchorHub will give information to the client about how to fix their
file(s) and exit.